+++ sdk/trace/tracetest/test.go
@@ -12,15 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package tracetest is a testing helper package for the SDK. User can configure no-op or in-memory exporters to verify
-// different SDK behaviors or custom instrumentation.
-package tracetest // import "go.opentelemetry.io/otel/sdk/export/trace/tracetest"
+// Package tracetest is a testing helper package for the SDK. User can
+// configure no-op or in-memory exporters to verify different SDK behaviors or
+// custom instrumentation.
+package tracetest // import "go.opentelemetry.io/otel/sdk/trace/tracetest"
 
 import (
 	"context"
 	"sync"
 
-	"go.opentelemetry.io/otel/sdk/export/trace"
+	"go.opentelemetry.io/otel/sdk/trace"
 )
 
 var _ trace.SpanExporter = (*NoopExporter)(nil)